refactor: move backdrop UI to Backdrop component
