+++ src/App.tsx
@@ -1,5 +1,5 @@
-import { useEffect, useState } from "react";
-import { NoteMessageEvent, WebMidi } from "webmidi";
+import { useState } from "react";
+// import { NoteMessageEvent, WebMidi } from "webmidi";
 
 import { Controls } from "./Controls";
 import { Overlay } from "./Overlay";
@@ -29,97 +29,59 @@ const keys = [
   "'",
 ];
 
-const keyIndexToHue = (index: number) => Math.floor((index * 360) / 12);
-
-const hueToColorString = (hue: number) => `hsl(${hue} 60% 50%)`;
-
-const huesToLinearGradient = (
-  hues: Array<number>,
-  angleInDeg: number,
-  xPos: number,
-  yPos: number
-) =>
-  `conic-gradient( from ${angleInDeg}deg at ${xPos}% ${yPos}%, ${hues
-    .map(hueToColorString)
-    .join(",")})`;
-
 export default function App() {
   const [keysPressed, setKeysPressed] = useState<Array<number>>([]);
   const [keyCharsPressed, setKeyCharsPressed] = useState<Array<string>>([]);
-  const [background, setBackground] = useState<string>();
-  const [opacity, setOpacity] = useState(0);
-  const [angleInDeg, setAngleInDeg] = useState<number>(
-    Math.floor(Math.random() * 360)
-  );
-  const [xPos, setXPos] = useState<number>(Math.random() * 100);
-  const [yPos, setYPos] = useState<number>(Math.random() * 100);
-  const [initialized, setInitialized] = useState(false);
+
+  // const [initialized, setInitialized] = useState(false);
 
   const [isMuted, setMuted] = useState(false);
 
-  const onNoteOn = (e: NoteMessageEvent) => {
-    const key = e.note.number - 60;
-    setKeysPressed((keysPressed) => {
-      if (keysPressed.includes(key)) {
-        return keysPressed;
-      }
-      const newKeysPressed = keysPressed.slice();
-      newKeysPressed.push(key);
-      return newKeysPressed;
-    });
-  };
+  // const onNoteOn = (e: NoteMessageEvent) => {
+  //   const key = e.note.number - 60;
+  //   setKeysPressed((keysPressed) => {
+  //     if (keysPressed.includes(key)) {
+  //       return keysPressed;
+  //     }
+  //     const newKeysPressed = keysPressed.slice();
+  //     newKeysPressed.push(key);
+  //     return newKeysPressed;
+  //   });
+  // };
 
-  const onNoteOff = (e: NoteMessageEvent) => {
-    const key = e.note.number - 60;
-    setKeysPressed((keysPressed) => {
-      const newKeysPressed = keysPressed.slice();
-      const index = newKeysPressed.indexOf(key);
-      if (index > -1) {
-        newKeysPressed.splice(index, 1);
-      }
-      return newKeysPressed;
-    });
-  };
+  // const onNoteOff = (e: NoteMessageEvent) => {
+  //   const key = e.note.number - 60;
+  //   setKeysPressed((keysPressed) => {
+  //     const newKeysPressed = keysPressed.slice();
+  //     const index = newKeysPressed.indexOf(key);
+  //     if (index > -1) {
+  //       newKeysPressed.splice(index, 1);
+  //     }
+  //     return newKeysPressed;
+  //   });
+  // };
 
-  useEffect(() => {
-    WebMidi.enable()
-      .then(() => {
-        setInitialized(true);
-      })
-      .catch((err) => console.log(err));
-  }, []);
+  // useEffect(() => {
+  //   WebMidi.enable()
+  //     .then(() => {
+  //       setInitialized(true);
+  //     })
+  //     .catch((err) => console.log(err));
+  // }, []);
 
-  useEffect(() => {
-    if (!initialized) {
-      return;
-    }
-    WebMidi.inputs.forEach((input) => {
-      input.addListener("noteon", onNoteOn);
-      input.addListener("noteoff", onNoteOff);
-    });
-  }, [initialized]);
-
-  useEffect(() => {
-    if (keysPressed.length) {
-      const hues = keysPressed.map(keyIndexToHue);
-      setOpacity(1);
-      if (hues.length > 1) {
-        setBackground(huesToLinearGradient(hues, angleInDeg, xPos, yPos));
-      } else {
-        setBackground(hueToColorString(hues[0]));
-      }
-    } else {
-      setOpacity(0);
-      setAngleInDeg(Math.floor(Math.random() * 360));
-      setXPos(Math.random() * 100);
-      setYPos(Math.random() * 100);
-    }
-  }, [keysPressed, angleInDeg, xPos, yPos]);
+  // useEffect(() => {
+  //   if (!initialized) {
+  //     return;
+  //   }
+  //   WebMidi.inputs.forEach((input) => {
+  //     input.addListener("noteon", onNoteOn);
+  //     input.addListener("noteoff", onNoteOff);
+  //   });
+  // }, [initialized]);
 
   window.onkeydown = (e) => {
     const keyChar = e.key;
 
-    console.log("keydown)");
     const index = keys.indexOf(keyChar);
     if (keysPressed.includes(index) || index < 0) {
       return;
@@ -164,7 +126,7 @@ export default function App() {
 
   return (
     <div className="app">
-      <Backdrop {...{ background, opacity }} />
+      <Backdrop keysPressed={keysPressed} />
       <SynthConnector {...{ keysPressed }} isMuted={isMuted} />
       <Overlay {...{ keyCharsPressed }} />
       <Controls isMuted={isMuted} toggleMuted={() => setMuted(!isMuted)} />
